authapp/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
defusedxml==0.7.1
Django==4.0
django-countries==7.2.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
httpie==2.6.0
idna==3.3
Naked==0.1.31
Pillow==8.4.0
psycopg2-binary==2.9.2
pycrypto==2.6.1
Pygments==2.10.0
PyJWT==2.3.0
PySocks==1.7.1
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-toolbelt==0.9.1
sqlparse==0.4.2
urllib3==1.26.7
gunicorn==20.0.4
whitenoise==5.2.0